Fixed what I think is an error in FileCache. (#922)
The private method always zeroed the cache size information
in the cache info file (which doubles as a cache lock file).
I modified the code to look at the size of the cache info
file and if it's zero length, initialize it to zero. Otherwise,
do not set it to zero.