Cmr lpdaac patch (#928)
Added a patch and two tests for LPDAAC's unusual use of the RelatedURL information

In CMR, LPDAAC stores an HTTPS/S3 pair of URLs to a file that is apparently
XML metadata from or for CMR (the file ends in '.cmr.xml'). This was tripping 
up the NGAP API which returned that URL as the URL to the data granule. I
added to heuristic that looks for the data URL so that it won't return these XML
file URLs. 

Also, I modified NgapApiTest to use the common unit test code and added
two tests for the function that encodes the heuristic lookup process. Each is
based on 'canned' JSON from CMR so the tests don't depend on CMR running,
et cetera.