An error message was not actually an error.
I changed the message for FileCache::put so that when two
put calls collide, but only one adds the given key, there
is an IFO message written and not an error. The cache is
working correctly - the items is added only once - and
changing the caller to eliminate this for the small number
of cases where it happens would be worse than just letting
it happen. In one sample, this happened 265 times in 22392
calls to put (1.2%)