Fixed glaring errors in the NgapOwnedContainer that made it to master (#962)
Fixed glaring errors in the NgapOwnedContainer that made it to master

Tested (manually) getting a DMR++ from a DAAC after trying OPeNDAP.
That is, a granule's DMR++ is not in the OPeNDAP-owned bucket,
but is in the DAAC bucket. If we try the OPeNDAP bucket and don't
find the granule, does it then work to get the DMR++ from the
DAAC's bucket? Yes.
Also added Sonar fixes and a README about the integration tests.