Dmrpp bucket option (#963)
Fixed glaring errors in the NgapOwnedContainer that made it to master

Tested (manually) getting a DMR++ from a DAAC after trying OPeNDAP

That is, a granule's DMR++ is not in the OPeNDAP-owned bucket,
but is in the DAAC bucket. If we try the OPeNDAP bucket and don't
find the granule, does it then work to get the DMR++ from the
DAAC's bucket? Yes.

Updates to the ngap tests that don't run on Travis
These require credentials that are not avaialble with Travis.

Started adding 'trust' dmrpp attribute
Somewhat less tortured logic in get_dmrpp_from_cache_or_remote_source()
Addressed a complexity issue in get_dmrpp_from_cache_or_remote_source()