More C++14 refactoring and a bug fix.
The SuperChunkTest and DmrppArrayTest unit tests failed on the M1
because mutexes in CurlHandlePool were being missued in the unit
tests. The DmrppRequestHandler was initializing and clearing the
curl library, but not the CurlHandlePool, and this invalidated mutexes
that the CH Pool used with the curl library (see libcurl's share interface).
Fixed by changing the static unique_ptr to a plain pointer and managing
its lifetime the old way.