Fix the info message issue and write tests (#1005)
Change how we evaluate URLs for S3 signing

Fix the info messages in CurlHandlePool. These are too chatty.
Also refactored code to use CurlUtils function for S3 signing and
remove a use of strlen where its safety was dubious.

The SuperChunkTest and DmrppArrayTest unit tests failed on the M1
because mutexes in CurlHandlePool were being misused in the unit
tests. The DmrppRequestHandler was initializing and clearing the
curl library, but not the CurlHandlePool, and this invalidated mutexes
that the CH Pool used with the curl library (see libcurl's share interface).
Fixed by changing the static unique_ptr to a plain pointer and managing
its lifetime the old way.