Added a compile-time switch for the fcntl(2) code and fixed put()
The issue with put(), where it takes the FileCache::PutItem object
and thus transfers control of the Item lock back to the caller, is
that the cache_info file is modified by the PutItem dtor without
locking the cache. Two processes can update the cache_info file at
the same time. Not good.