Fixed the issue with put() - both versions were broken.
Summary: On a single machine when one process opens a file and then
forks, the parent and child processs have the same ddescriptor open to
the file. flock(2) will not work in this case. The child must close
the descriptor from the parent and open the file on its own, so that
it has its own descriptor to the open file. Once done, flock(2) works
as expected. I used flock(2) instead of fcntl(2) because flock(2) is
thread safe, which saves on bookkeeping within a process.