Test the proposed fix for FileCache::put(const string&, PutItem&) (#1028)
Fixed a bug with put() - both versions were corrupting the cache_info file

The issue with put(), where it takes the FileCache::PutItem object
and thus transfers control of the Item lock back to the caller, is
that the cache_info file is modified by the PutItem dtor without
locking the cache. Two processes can update the cache_info file at
the same time. Not good.

A 'feature' of flock(2):  On a single machine when one process opens a file and then
forks, the parent and child processes have the same descriptor open to
the file. flock(2) will not work in this case. The child must close
the descriptor from the parent and open the file on its own, so that
it has its own descriptor to the open file. Once done, flock(2) works
as expected. I used flock(2) instead of fcntl(2) because flock(2) is
thread safe, which saves on bookkeeping within a process.

Speed them up and, ensure struct stat variables are initialized and
make sure all fork(2) uses properly open their own copy of the
FileCache object.