Fixed a distcheck fail in the FileCacheTest code.
The read_file() function did not signal an error well, so when
I forgot to include two files in teh EXTRA_DIST variable, there
was no failure when the test tried to open the missing files.
There was just an odd, incorrect value, which made the test fail
for an inexplicable reason.