Fix for the DAP Function Result cache inconsistency bug.
This cache inconsistency bug was caused when a constraint
expression function failed and the to-be-cached response
was thusly corrupted. The code tried to remove that response
file from the cache, but failed to remove both the initial
lock on the file and the second lock added by the 'purge_file()'
method of BESFileLockingCache. This bug may have been at
the root of other BES FLC errors since it would be triggered
by _any_ code that threw an exception when BES FLC tried to
cache a file.