[HYRAX-1891] Test AWS C++ SDK pre-signed url behavior from bes (#1161)
* WIP

* Added AWS SDK code for S3 GET and HEAD operations.

This works with S3 objects that must be signed. It
reads teh AWS Key and Secret from environment variables.

* Moving the HttpHead code - likely will not be used.

* Added AWS_SDK class

* Added an interface class for AWS_SDK.cc

This will enable using the dependency injection pattern
so that we can write unit-tests for users of AWS_SDK
without actually having to call a web server.

* Added unit tests for the AWS_SDK class

Also added HTTP response code and error message access to the class

* Refactoring plus NgapOwnedContainer test patch

There was a test I added for S3 access that was failing. I
marked it as such. This will be fixed soon.

* WIP

* WIP - Fix CredentialsManagerTests

* Re-introduced CredentialsManagerTest with a temp patch.

* WIP: Add the aws sdk as a separate convienence library

autoconf finds the headers but not the libraries and the
unit-tests need work.

* Repair framework-only build when AWS SDK is not present

* Added test_config.h.in to AWS_SDK unit tests

* WIP - switch from libcurl to the AWS SDK

* Git test updates - the basic idea of the container has shifted somewhat.

Now the container works for either a DAAC URL (https://...) or
an object and bucket name pair. The file URLs that used to work
are no longer supported.

* Modify configure to add -L$prefix/deps/lib for the AWS SDK

This is a WIP

* WIP: working on libbes_dispatch link path conflict w/the AWS libs

* Fixes for M4 and missing env vars inteh AWS unit tests

* Added bes dispatch lib to AWS_SDK_Test link

* Fixed AWS tests when the CMAC evn vars are not set

* WIP: Try to reimplement the signing code.

This is a start down that path.

* Back the AWS SDK out of the code for now. WIP

* undo unrelated changes

* Undo more changes

* remove more

* undo http changes

* undo more

* remove commented out

* remove commented out code in aws_sdk.h

* remove duplication in configure.ac, fail if aws sdk not found

* rename for legibility

* clean up

* make aws dependency required not optional

* Find the AWS C++ SDK in the hyrax-dependencies.

This is a bit lame since it assumes that if the
AWS headers are present in deps/include, then the
libraries are there in lib. Probably good enough,
though.

* renaming for clarity

* add presign function

* add basic unit tests for presigning

* add test that signed url object can be used to fetch object

* test signed url expiration

* Update aws/AWS_SDK.cc

* merge

* Fixed an error in s3_reader's Makefile.am

'unit-tests' was added to SUBDIRS, but the Makefile there is not built.

* AWS SDK unit test linkage fixed on OSX

---------

Co-authored-by: James Gallagher <jgallagher@opendap.org>
Co-authored-by: Hannah Robertson <1076701+hannahilea@users.noreply.github.com>