[HYRAX-1900] Various AWS/S3 fixs (#1171)
Changes to use CMAC_ON=yes|no to control tests that use our S3 bucket.
If the env var CMAC_ON is defined and set to 'yes,' then the other CMAC_*
env vars are used and tests that require authenticated access to S3 will be
run. If the env var is not set to 'yes' or not defined, the tests are not run.
Co-authored-by: Hannah Robertson <hannahilea@users.noreply.github.com>