I am trying Ubuntu jammy to get Java 17
This will stop us from having to load java 17 as a separate step,
which is often the cause for failed CICD builds. Also, I
have set CMAC_ON=yes which is needed after a recent branch
merge that introduced this feature. The CICD builds _do_
run the S3 tests, so that switch should be set.