HYRAX-1898 cmr client id header added (#1198)
Changed from query string to headers. Needs tests.
NgapApiTest updated because the CMR Client Id mechanism changed

We now use a request header and not a query parameter. Because of
that, the test for this is in bes/http. That implementation should
be here, but that is a bigger refactor.