HYRAX-1953 jhrg/bes warnings (#1196)
Many fixes for warnings in bes/dispatch and bes/dispatch/unit-tests.
Added README for clang-tidy (in bes/) and a pair of scripts that apply
various 'modernizations' and reformat code, using clang-tidy. The 
scripts are in bes/dispatch and are named 'cpp-... .sh'. We might work
on those more and move them up to the top, but this process of 
modernizing the code is tedious and requires lots of iteration: make
a change, test, make another change, ...