The distcheck target is broken on master; Fix (#1210)
Code in scale_util.cc used the old Array::size() method.
I replaced that call with a call to Array::length().

Added a way to toggle GDAL builds for the BES.
This operates on the Travis builds only and is
independent of the presence of gdal in the hyrax
dependencies build.

Fixed some 'code smells' but I was pretty conservative 
with them, since this code is probably not going to be built
often until the issues with GDAL linkage are worked out.